Animation system is functional!
Default animation runs at 2 frames per second, this will probably be changed to require an actual frame input. Animation speed can be overwritten.

Animation can also be mirrored (this works great for the stick figure testing character but will probably not be useful for the actual player character if the character is not symetrical.